refactor: upbitAdapter DI 대신 직접 의존성 주입
